+++ Plugins/iOS/CarnivalWrapper.h
@@ -28,7 +28,7 @@ extern "C" void _setString(const char *string, const char *key);
 extern "C" void _setBool(bool boolValue, const char *key);
 extern "C" void _setDate(int64_t secondsSince1970, const char *key);
 extern "C" void _setFloat(float floatValue, const char *key);
-extern "C" void _setInteger(int64_t intValue, const char *key);
+extern "C" void _setInteger(int intValue, const char *key);
 extern "C" void _removeAttribute(const char *key);
 
 //Device ID
